fix(kernel): round-6 review — int parity, port/protocol reject, untag safety classifier, docs
- Narrow ints (M1-r6): kernel ScanCell returned int64 for TINYINT/SMALLINT/INT
  while Thrift returns native int8/int16/int32, so a top-level narrow int differed
  by Go type across backends (verified live). Return the native width to match
  Thrift (its prod callers make widening Thrift the breaking option; the kernel has
  none). Both are technically off the driver.Value spec (int64 only); unifying both
  on int64 is a tracked driver-wide follow-up (needs team sign-off — see
  decisions-and-deferrals.md). Parity test gains tinyint/smallint/int; e2e int
  expectations updated to native widths.
- Port/Protocol (M2-r6): the drop-guard classified both "forwarded", but
  kernel.Config has neither field — the kernel takes a bare host and connects over
  https:443. So a non-default WithPort / non-https protocol was silently ignored on
  the kernel path (the exact divergence the "nothing silently ignored" contract
  forbids). Reject a non-default port/protocol in validateKernelConfig; reclassify
  as "rejected". Defaults (https/443) still validate.
- Safety classifier in CI (M3-r6): isBadConnection/isSessionFatal/toStatementError/
  toConnError + KernelError + the status consts enforce the H1 (no statement replay)
  and L1 (no permanent-auth retry) guarantees, but lived behind the cgo build tag
  and ran in no CI job. Move them to an untagged errors_classify.go (plain-int
  consts with compile-time lockstep assertions against the C enum in cgo.go); move
  their tests to an untagged errors_classify_test.go so they run under
  CGO_ENABLED=0. A regression now fails default CI instead of shipping green.
- Docs (L1-r6): document useKernel/warehouseId DSN params + WithUseKernel/
  WithWarehouseID options in doc.go.

Verified: default suite + golangci-lint clean; kernel-tagged unit; live e2e +
Thrift-parity against staging (all int widths and rejects confirmed).
